Import levels on level load instead of game load.
This PR refactors code so that levels are imported on level load instead of game load. It does this by getting rid of the Level struct and using the ImportRequest struct to dynamically load level resources on load. This allows memory to be allocated when its needed, and immediately dumped when it is not. Unfortunately, I had to re-add IniReader as a dependency of LevelImporter, but I think the trade-off is well worth it.